Move extract_facade_areas to shared_tasks
